refactor: dissolve domain/services into domain/plan
The differ lived in domain/services/, a folder that earned nothing: services is
a non-name (a role in an OOP pattern, not a domain concept), its __init__ was
empty (no curated surface), and it wrapped a single file. Callers reached it by
full path, domain.services.differ -- the folder was pure ceremony.

The differ PRODUCES an ActionPlan and imports eight names from plan/actions.py;
it changes together with the action vocabulary every time an action kind is
added. So its real home is domain/plan/. Moving differ.py there:

  domain/plan/
    __init__.py   # unchanged: still exports only the action TYPE surface
    actions.py    # the change vocabulary (what can happen)
    differ.py     # how to compute the change (was services/differ.py)

This also graduates plan/ from a single-file package (its own bit of ceremony)
into a genuine two-file package that earns its existence. The curated __init__
keeps exporting only the type surface; callers wanting the computation import
domain.plan.differ by full path, keeping the plan artefact separate from how it
is built.

model/ was reviewed and deliberately left intact: its four files (column,
data_type, qualified_name, table) change on independent axes, so merging any of
them would couple orthogonal growth.

Pure move plus import rewrite, no logic change: differ.py and its tests
relocate, two import lines update (application/plan.py, test_differ.py), the
empty services __init__ files are deleted. Full unit suite passes unchanged;
lint and mypy clean. Future domain concepts stay additive (a new concept adds a
model file, action subclasses, and a differ helper -- no re-foldering).
