+++ src/delta_engine/domain/plan/differ.py
